# docs/dialoga-openapi.yaml
openapi: 3.0.3
info:
  title: Dialoga API
  description: 'API REST do backend da plataforma Dialoga (monitoramento terapêutico), cobrindo os perfis
    Paciente, Profissional e Administrador.


    Gerado a partir do modelo lógico `BD-lógico-Dialoga_2026.mwb` e validado contra os 16 documentos de
    Requisitos Funcionais (RF_2026_001 a RF_2026_016). Cada operação referencia os IDs de RF que originaram
    a rota, para rastreabilidade acadêmica.'
    
  version: 1.0.0
  contact:
    name: Dialoga — Projeto Profissional de Software, SENAI Jandira
servers:
- url: https://api.dialoga.com.br/v1
  description: Produção (placeholder — ajustar quando o backend for publicado)
...
paths:
  /auth/login:
    post:
      summary: Login por e-mail e senha
      description: 'Requisitos relacionados: RF85'
      tags:
      - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    $ref: '#/components/schemas/LoginResponse'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
  /auth/login/google:
    post:
      summary: Login via Google OAuth
      description: 'Requisitos relacionados: RF86 — TODO: requer coluna de senha opcional no modelo (ver
        observações)'
      tags:
      - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginGoogleRequest'
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    $ref: '#/components/schemas/LoginResponse'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
...
  /auth/recuperar-senha/solicitar:
    post:
      summary: Solicitar código de recuperação de senha
      description: 'Requisitos relacionados: RF87, RF88, RF92, RF93'
      tags:
      - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecuperarSenhaSolicitarRequest'
      responses:
        '200':
          description: Código enviado por e-mail
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Código enviado por e-mail
                  response:
                    type: object
                    nullable: true
                    example: {}
                required:
                - status
                - status_code
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
  /auth/recuperar-senha/validar-codigo:
    post:
      summary: Validar código de recuperação recebido por e-mail
      description: 'Requisitos relacionados: RF89, RF94, RF95'
      tags:
      - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecuperarSenhaValidarCodigoRequest'
      responses:
        '200':
          description: Código válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Código válido
                  response:
                    type: object
                    nullable: true
                    example: {}
                required:
                - status
                - status_code
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
  /auth/recuperar-senha/redefinir:
    post:
      summary: Redefinir senha após validação do código
      description: 'Requisitos relacionados: RF90, RF96, RF97'
      tags:
      - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecuperarSenhaRedefinirRequest'
      responses:
        '200':
          description: Senha redefinida
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Senha redefinida
                  response:
                    type: object
                    nullable: true
                    example: {}
                required:
                - status
                - status_code
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
  /pacientes:
    post:
      summary: Cadastrar novo paciente
      description: 'Requisitos relacionados: RF98-106'
      tags:
      - Pacientes
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PacienteCadastroRequest'
      responses:
        '201':
          description: Paciente criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: Registro inserido com sucesso!
                  response:
                    $ref: '#/components/schemas/Paciente'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
...
  /pacientes/me:
    get:
      summary: Consultar próprio perfil de paciente
      description: 'Requisitos relacionados: RF153'
      tags:
      - Pacientes
      responses:
        '200':
          description: Perfil do paciente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    $ref: '#/components/schemas/Paciente'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Editar próprio perfil de paciente
      description: 'Requisitos relacionados: RF154, RF157-159'
      tags:
      - Pacientes
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PacienteAtualizarRequest'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Registro atualizado com sucesso!
                  response:
                    $ref: '#/components/schemas/Paciente'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /profissionais/me/pacientes:
    get:
      summary: Listar pacientes vinculados ao profissional autenticado
      description: 'Requisitos relacionados: RF299-301, RF306'
      tags:
      - Vínculos
      parameters:
      - name: busca
        in: query
        required: false
        schema:
          type: string
        description: Busca por nome (RF300)
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - Todos
          - Ativos
          - Inativos
        description: Filtro de status (RF301)
      responses:
        '200':
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: array
                    items:
                      $ref: '#/components/schemas/PacienteResumo'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Adicionar paciente (busca na plataforma ou cadastro externo)
      description: 'Requisitos relacionados: RF302-305'
      tags:
      - Vínculos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/VinculoCriarRequest'
              - $ref: '#/components/schemas/PacienteExternoCriarRequest'
      responses:
        '201':
          description: Vínculo criado ou paciente externo cadastrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: Registro inserido com sucesso!
                  response:
                    $ref: '#/components/schemas/Vinculo'
                required:
                - status
                - status_code
                - response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profissionais/me/pacientes/{vinculoId}:
    delete:
      summary: Remover paciente da lista de acompanhamento
      description: 'Requisitos relacionados: RF315, RF326-330'
      tags:
      - Vínculos
      parameters:
      - name: vinculoId
        in: path
        required: true
        description: Identificador do vínculo
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Vínculo movido para histórico (status Inativo)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Registro apagado com sucesso!
                  response:
                    type: object
                    nullable: true
                    example: {}
                required:
                - status
                - status_code
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Erro:
      type: object
      properties:
        status:
          type: boolean
          example: false
        status_code:
          type: integer
          example: 400
        message:
          type: string
          example: Os dados enviados na requisição não estão corretos.
      required:
      - status
      - status_code
      - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
          format: password
          minLength: 6
          maxLength: 255
      required:
      - email
      - senha
    LoginGoogleRequest:
      type: object
      properties:
        id_token_google:
          type: string
          description: ID token retornado pelo OAuth do Google (RF86).
      required:
      - id_token_google
    LoginResponse:
      type: object
      properties:
        token_acesso:
          type: string
        tipo_perfil:
          type: string
          enum:
          - paciente
          - profissional
          - administrador
        id_usuario:
          type: integer
          format: int64
    RecuperarSenhaSolicitarRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
      - email
    RecuperarSenhaValidarCodigoRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        codigo:
          type: string
          minLength: 4
          maxLength: 4
          description: Código numérico de 4 dígitos (RF88/RF94).
      required:
      - email
      - codigo
    RecuperarSenhaRedefinirRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        codigo:
          type: string
        nova_senha:
          type: string
          format: password
          minLength: 6
          maxLength: 255
      required:
      - email
      - codigo
      - nova_senha
...
    Paciente:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome_completo:
          type: string
          maxLength: 150
        apelido:
          type: string
          maxLength: 80
          nullable: true
        data_nascimento:
          type: string
          format: date
          nullable: true
        celular:
          type: string
          maxLength: 20
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        foto_avatar:
          type: string
          nullable: true
        deseja_iniciar_registros:
          type: boolean
        primeiro_acesso_concluido:
          type: boolean
        status_atividade:
          type: string
          maxLength: 20
          example: Ativo
        criado_em:
          type: string
          format: date-time
          nullable: true
    PacienteResumo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome_completo:
          type: string
        foto_avatar:
          type: string
          nullable: true
        status_atividade:
          type: string
    PacienteCadastroRequest:
      type: object
      properties:
        nome_completo:
          type: string
          minLength: 3
          maxLength: 255
        celular:
          type: string
          pattern: ^[0-9]{11}$
        email:
          type: string
          format: email
          minLength: 10
          maxLength: 255
        senha:
          type: string
          format: password
          minLength: 6
          maxLength: 255
        foto_avatar:
          type: string
          format: binary
          nullable: true
          description: Arquivo de imagem do avatar (multipart/form-data), opcional no cadastro.
        aceite_termos:
          type: boolean
      required:
      - nome_completo
      - email
      - senha
      - aceite_termos
...
    PacienteAtualizarRequest:
      type: object
      properties:
        nome_completo:
          type: string
        apelido:
          type: string
        celular:
          type: string
        foto_avatar:
          type: string
          format: binary
          nullable: true
          description: Arquivo de imagem do avatar (multipart/form-data).
    PacienteExternoCriarRequest:
      type: object
      description: Cadastro leve de paciente feito pelo profissional, para paciente sem conta na plataforma
        (RF304). Requer ajuste de modelo — ver observações no topo do documento.
      properties:
        nome_completo:
          type: string
          minLength: 3
          maxLength: 255
      required:
      - nome_completo
...
    ProfissionalResumo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome_completo:
          type: string
        especialidade:
          type: string
          nullable: true
        instituicao_clinica:
          type: string
          nullable: true
...
    Vinculo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - Pendente
          - Aceito
          - Recusado
          - Inativo
        origem_cadastro:
          type: string
          description: 'Quem originou o vínculo, ex.: ''Paciente'' ou ''Profissional''.'
          maxLength: 20
        data_solicitacao:
          type: string
          format: date-time
          nullable: true
        data_aceite:
          type: string
          format: date-time
          nullable: true
        data_inicio_acompanhamento:
          type: string
          format: date-time
          nullable: true
        data_fim:
          type: string
          format: date-time
          nullable: true
        paciente:
          $ref: '#/components/schemas/PacienteResumo'
        profissional:
          $ref: '#/components/schemas/ProfissionalResumo'
    VinculoCriarRequest:
      type: object
      description: Solicita vínculo entre paciente e profissional (RF172-176, RF302-305).
      properties:
        paciente_id:
          type: integer
          format: int64
        profissional_id:
          type: integer
          format: int64
      required:
      - paciente_id
      - profissional_id
...
  responses:
    BadRequest:
      description: Os dados enviados na requisição não estão corretos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 400
            message: Os dados enviados na requisição não estão corretos.
    Unauthorized:
      description: Não foi possivel processar requisição. usuário não autorizado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 401
            message: Não foi possivel processar requisição. usuário não autorizado.
    NotFound:
      description: Não foi possivel encontrar nenhum registro.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 404
            message: Não foi possivel encontrar nenhum registro.
    Conflict:
      description: Não foi possivel processar requisição. usuario ou email já cadastrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 409
            message: Não foi possivel processar requisição. usuario ou email já cadastrado.
    UnsupportedMediaType:
      description: Não foi possivel processar a requisição, pois o formato de dados aceito pela API é
        somente JSON.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 415
            message: Não foi possivel processar a requisição, pois o formato de dados aceito pela API
              é somente JSON.
    InternalServerError:
      description: Não foi possivel processar a requisição por conta de erro interno na API.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            status: false
            status_code: 500
            message: Não foi possivel processar a requisição por conta de erro interno na API.
security:
- bearerAuth: []